refactored solution structure into folders for maps, squares and tiles.
