+++ Robot/Squares/Square.cs
@@ -1,6 +1,7 @@
-﻿using System;
+﻿using Game;
+using Tiles;
 
-namespace Game
+namespace Squares
 {
     public abstract class Square
     {
@@ -109,62 +110,4 @@ namespace Game
             return neighbor;    // can be null if off the map.
         }
     }
-
-    public class GrassSquare : Square
-    {
-        public override Tile Tile { get { return FieldTileSet.Grass; } }
-        public override string Name { get { return "Land"; } }
-        public override bool Passable { get { return true; } }
-
-        public GrassSquare(int x, int y) : base(x, y) { }
-    }
-
-    public class WaterSquare : Square
-    {
-        public override Tile Tile { get { return FieldTileSet.Water; } }
-        public override string Name { get { return "Water"; } }
-        public override bool Passable { get { return false; } }
-
-        public WaterSquare(int x, int y) : base(x, y) { }
-    }
-
-    public class TreeSquare : Square
-    {
-        public override Tile Tile { get { return FieldTileSet.Tree; } }
-        public override string Name { get { return "Tree"; } }
-        public override bool Passable { get { return false; } }
-
-        public TreeSquare(int x, int y) : base(x, y) { }
-    }
-
-    public class DoorSquare : Square
-    {
-        public override Tile Tile { get { return DoorTileSet.Closed; } }
-        public override string Name { get { return "Door"; } }
-        public override bool Passable { get { return true; } }
-
-        private string EventName { get; }
-
-        public DoorSquare(int x, int y, string eventName) : base(x, y) 
-        {
-            EventName = eventName;
-        }
-
-        public override bool EnterSquare(Player player)
-        {
-            bool moved = base.EnterSquare(player);
-
-            if (moved)
-            {
-                GameEvents.GetGameEvents()
-                    .AddEvent(new GameEvent() { 
-                        Player = player, 
-                        Type = GameEvent.EventType.ENTERED_DOOR,
-                        Name = EventName});
-            }
-
-            return moved;
-        }
-    }
-
 }
