+++ Robot/Tiles/RobotTileSet.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Game;
 
-namespace Game
+namespace Tiles
 {
     public static class RobotTileSet 
     {
